move charts
