+++ vue/src/charts/README.md
@@ -1,6 +1,6 @@
 # 这里提供了 SwanLab 所有支持的图表类型
 
-每个图表类型被抽象为一个组件，在 [ChartContainer](../ChartContainer.vue) 中选择性渲染。
+每个图表类型被抽象为一个组件，在其他地方使用时，使用component标签选择性渲染
 
 ## 图表组件
 
@@ -13,7 +13,7 @@
 | chart | 图表配置 |
 | title | 图表标题 |
 
-每个图表组件应该暴露一些通用 api，在[ChartContainer](../ChartContainer.vue)中通过 ref 获取组件对象进行调用，具体如下：
+每个图表组件应该暴露一些通用 api，在其他地方通过 ref 获取组件对象进行调用，具体如下：
 
 | api    | 具体作用                       |
 | ------ | ------------------------------ |
